+++ RemoteClient/Tool.h
@@ -39,7 +39,7 @@ public:
 		if (hRet == S_OK)
 		{
 			ULONG length = 0;
-			pStream->Write(pData, strBuffer.size(), &length);
+			pStream->Write(pData, (ULONG)strBuffer.size(), &length);
 			LARGE_INTEGER bg = { 0 };
 			pStream->Seek(bg, STREAM_SEEK_SET, NULL);
 			if (image.IsDIBSection())
